requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
pymongo==4.5.0
python-dotenv==0.19.0
requests==2.26.0
opencv-python-headless==4.6.0.66
tensorflow-cpu==2.9.1
google-cloud-translate==3.7.0
openai==0.27.0
google-generativeai==0.3.0
twilio==7.3.0
python-jose==3.3.0
passlib==1.7.4
flask-cors==3.0.10
python-multipart==0.0.5
pillow==8.3.2
numpy==1.21.2
pandas==1.3.3
